api/requirements.txt
# AgencyZen API - Python Backend
# Requisitos para o backend

fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.12.0
replicate==0.25.0
pydantic==2.6.0
python-multipart==0.0.9
websockets==12.0
qrcode[pil]==7.4.2
aiosqlite==0.19.0
httpx==0.26.0
pillow==10.2.0
